# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GrantAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  orgName: string;
  orgType: string;
  mission: string;
  location: string;
  focusAreas: string;
  /** @nullable */
  fundingNeed?: number | null;
  createdAt: string;
}

export interface CreateUserBody {
  orgName: string;
  orgType: string;
  mission: string;
  location: string;
  focusAreas: string;
  /** @nullable */
  fundingNeed?: number | null;
}

export interface UpdateUserBody {
  orgName?: string;
  orgType?: string;
  mission?: string;
  location?: string;
  focusAreas?: string;
  /** @nullable */
  fundingNeed?: number | null;
}

export interface Grant {
  id: number;
  name: string;
  funder: string;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  deadline?: string | null;
  eligibility: string;
  focusArea: string;
  location: string;
  description: string;
  /** @nullable */
  website?: string | null;
  createdAt: string;
}

export interface FocusAreaCount {
  focusArea: string;
  count: number;
}

export interface GrantStats {
  totalGrants: number;
  totalFunding: number;
  byFocusArea: FocusAreaCount[];
  upcomingDeadlines: number;
}

export interface Match {
  userId: number;
  grantId: number;
  fitScore: number;
  /** @nullable */
  winProbability?: number | null;
  reasoning: string;
  createdAt: string;
}

export interface MatchWithGrant {
  userId: number;
  grantId: number;
  fitScore: number;
  /** @nullable */
  winProbability?: number | null;
  reasoning: string;
  createdAt: string;
  grant: Grant;
}

export interface GenerateMatchesBody {
  userId: number;
}

export interface Application {
  id: number;
  userId: number;
  grantId: number;
  /** @nullable */
  draftText?: string | null;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface ApplicationWithGrant {
  id: number;
  userId: number;
  grantId: number;
  /** @nullable */
  draftText?: string | null;
  status: string;
  createdAt: string;
  updatedAt: string;
  grant: Grant;
}

export interface CreateApplicationBody {
  userId: number;
  grantId: number;
  /** @nullable */
  draftText?: string | null;
  status?: string;
}

export interface UpdateApplicationBody {
  /** @nullable */
  draftText?: string | null;
  status?: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateAnthropicConversationBody {
  title: string;
}

export interface MagicModeBody {
  description: string;
}

export interface MagicModeMatchItem {
  grantId: number;
  fitScore: number;
  winProbability: number;
  reasoning: string;
  grant: Grant;
}

export interface MagicModeInferredProfile {
  orgName: string;
  mission: string;
  focusArea: string;
}

export interface MagicModeResult {
  inferredProfile: MagicModeInferredProfile;
  matches: MagicModeMatchItem[];
  draft: string;
  topGrantName: string;
  topGrantFunder: string;
}

export interface RewriteTextBody {
  selectedText: string;
  instruction: string;
}

export interface SendAnthropicMessageBody {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export type ListGrantsParams = {
  focusArea?: string;
  location?: string;
  search?: string;
};

export type ListMatchesParams = {
  userId: number;
};

export type ListApplicationsParams = {
  userId: number;
};