# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GrantAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /grants:
    get:
      operationId: listGrants
      tags: [grants]
      summary: List all grants
      parameters:
        - name: focusArea
          in: query
          required: false
          schema:
            type: string
        - name: location
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of grants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grant"
  /grants/{id}:
    get:
      operationId: getGrant
      tags: [grants]
      summary: Get grant details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Grant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationWithGrant"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create a new application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationBody"
      responses:
        "201":
          description: Created application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationWithGrant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application draft text or status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationBody"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /applications/{id}/generate:
    post:
      operationId: generateApplicationDraft
      tags: [ai]
      summary: Generate an AI draft for a grant application (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SSE stream of generated draft text
          content:
            text/event-stream: {}
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /applications/{id}/rewrite:
    post:
      operationId: rewriteApplicationText
      tags: [ai]
      summary: Rewrite a selected passage of application text (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewriteTextBody"
      responses:
        "200":
          description: SSE stream of rewritten text
          content:
            text/event-stream: {}
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Grant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        funder:
          type: string
        amount:
          type: ["integer", "null"]
        deadline:
          type: ["string", "null"]
        eligibility:
          type: string
        focusArea:
          type: string
        location:
          type: string
        description:
          type: string
        website:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - funder
        - eligibility
        - focusArea
        - location
        - description
        - createdAt
...
    Application:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        grantId:
          type: integer
        draftText:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - grantId
        - status
        - createdAt
        - updatedAt
    ApplicationWithGrant:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        grantId:
          type: integer
        draftText:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        grant:
          $ref: "#/components/schemas/Grant"
      required:
        - id
        - userId
        - grantId
        - status
        - createdAt
        - updatedAt
        - grant
    CreateApplicationBody:
      type: object
      properties:
        userId:
          type: integer
        grantId:
          type: integer
        draftText:
          type: ["string", "null"]
        status:
          type: string
      required:
        - userId
        - grantId
    UpdateApplicationBody:
      type: object
      properties:
        draftText:
          type: ["string", "null"]
        status:
          type: string
...
    RewriteTextBody:
      type: object
      properties:
        selectedText:
          type: string
        instruction:
          type: string
      required:
        - selectedText
        - instruction
...